+++ faker/faker_test.go
@@ -1,6 +1,8 @@
 package faker
 
 import (
+	"encoding/json"
+	"net"
 	"testing"
 	"time"
 
@@ -12,7 +14,8 @@ type testFakerStruct struct {
 	B string
 	C *string
 	D time.Time
-	E interface{}
+	E any
+	F json.Number
 }
 
 func TestFaker(t *testing.T) {
@@ -24,7 +27,8 @@ func TestFaker(t *testing.T) {
 	assert.NotEmpty(t, a.B)
 	assert.NotEmpty(t, a.C)
 	assert.NotEmpty(t, a.D)
-	assert.Empty(t, a.E) // empty interfaces are not faked
+	assert.Empty(t, a.E)                           // empty interfaces are not faked
+	assert.Equal(t, json.Number("123456789"), a.F) // json numbers should be numbers
 }
 
 type customType string
@@ -45,4 +49,35 @@ func TestFakerWithCustomType(t *testing.T) {
 	assert.NotEmpty(t, a.B)
 	assert.NotEmpty(t, a.C)
 	assert.NotEmpty(t, a.D)
+}
+
+type complexType struct {
+	IPAddress      net.IP
+	IPAddresses    []net.IP
+	PtrIPAddress   *net.IP
+	PtrIPAddresses []*net.IP
+	NestedComplex  struct {
+		IPAddress net.IP
+	}
+}
+
+func TestFakerCanFakeNetIP(t *testing.T) {
+	a := complexType{}
+	if err := FakeObject(&a); err != nil {
+		t.Fatal(err)
+	}
+	assert.NotEmpty(t, a.IPAddress)
+	assert.Equal(t, "1.1.1.1", a.IPAddress.String())
+
+	assert.Equal(t, 1, len(a.IPAddresses))
+	assert.Equal(t, "1.1.1.1", a.IPAddresses[0].String())
+
+	assert.NotEmpty(t, a.PtrIPAddress)
+	assert.Equal(t, "1.1.1.1", a.PtrIPAddress.String())
+
+	assert.Equal(t, 1, len(a.PtrIPAddresses))
+	assert.Equal(t, "1.1.1.1", a.PtrIPAddresses[0].String())
+
+	assert.NotEmpty(t, a.NestedComplex.IPAddress)
+	assert.Equal(t, "1.1.1.1", a.NestedComplex.IPAddress.String())
 }