+++ plugins/destination/plugin_test.go
@@ -1,4 +1,4 @@
-package plugins
+package destination
 
 import (
 	"context"
@@ -10,15 +10,15 @@ import (
 	"github.com/cloudquery/plugin-sdk/specs"
 )
 
-func TestDestinationPlugin(t *testing.T) {
-	p := NewDestinationPlugin("test", "development", NewTestDestinationMemDBClient)
-	DestinationPluginTestSuiteRunner(t, p, nil,
-		DestinationTestSuiteTests{})
+func TestPlugin(t *testing.T) {
+	p := NewPlugin("test", "development", NewTestDestinationMemDBClient)
+	PluginTestSuiteRunner(t, p, nil,
+		TestSuiteTests{})
 }
 
-func TestDestinationOnNewError(t *testing.T) {
+func TestOnNewError(t *testing.T) {
 	ctx := context.Background()
-	p := NewDestinationPlugin("test", "development", newTestDestinationMemDBClientErrOnNew)
+	p := NewPlugin("test", "development", newTestDestinationMemDBClientErrOnNew)
 	err := p.Init(ctx, getTestLogger(t), specs.Destination{})
 
 	if err == nil {
@@ -26,10 +26,10 @@ func TestDestinationOnNewError(t *testing.T) {
 	}
 }
 
-func TestDestinationOnWriteError(t *testing.T) {
+func TestOnWriteError(t *testing.T) {
 	ctx := context.Background()
 	newClientFunc := getNewTestDestinationMemDBClient(withErrOnWrite())
-	p := NewDestinationPlugin("test", "development", newClientFunc)
+	p := NewPlugin("test", "development", newClientFunc)
 	if err := p.Init(ctx, getTestLogger(t), specs.Destination{}); err != nil {
 		t.Fatal(err)
 	}
@@ -53,10 +53,10 @@ func TestDestinationOnWriteError(t *testing.T) {
 	}
 }
 
-func TestDestinationOnWriteCtxCancelled(t *testing.T) {
+func TestOnWriteCtxCancelled(t *testing.T) {
 	ctx := context.Background()
 	newClientFunc := getNewTestDestinationMemDBClient(withBlockingWrite())
-	p := NewDestinationPlugin("test", "development", newClientFunc)
+	p := NewPlugin("test", "development", newClientFunc)
 	if err := p.Init(ctx, getTestLogger(t), specs.Destination{}); err != nil {
 		t.Fatal(err)
 	}