+++ plugins/source/docs.go
@@ -1,27 +1,33 @@
-package plugins
+package source
 
 import (
+	"bytes"
 	"embed"
 	"encoding/json"
 	"fmt"
 	"os"
 	"path/filepath"
+	"regexp"
 	"strings"
 	"text/template"
 
+	"github.com/cloudquery/plugin-sdk/plugins/destination"
 	"github.com/cloudquery/plugin-sdk/schema"
 )
 
 //go:embed templates/*.go.tpl
 var templatesFS embed.FS
 
-// GenerateSourcePluginDocs creates table documentation for the source plugin based on its list of tables
-func (p *SourcePlugin) GenerateSourcePluginDocs(dir, format string) error {
+var reMatchNewlines = regexp.MustCompile(`\n{3,}`)
+var reMatchHeaders = regexp.MustCompile(`(#{1,6}.+)\n+`)
+
+// GeneratePluginDocs creates table documentation for the source plugin based on its list of tables
+func (p *Plugin) GeneratePluginDocs(dir, format string) error {
 	if err := os.MkdirAll(dir, os.ModePerm); err != nil {
 		return err
 	}
 
-	SetDestinationManagedCqColumns(p.Tables())
+	destination.SetDestinationManagedCqColumns(p.Tables())
 
 	switch format {
 	case "markdown":
@@ -46,7 +52,7 @@ type jsonColumn struct {
 	IsPrimaryKey bool   `json:"is_primary_key,omitempty"`
 }
 
-func (p *SourcePlugin) renderTablesAsJSON(dir string) error {
+func (p *Plugin) renderTablesAsJSON(dir string) error {
 	tables := p.jsonifyTables(p.Tables())
 	b, err := json.MarshalIndent(tables, "", "  ")
 	if err != nil {
@@ -56,7 +62,7 @@ func (p *SourcePlugin) renderTablesAsJSON(dir string) error {
 	return os.WriteFile(outputPath, b, 0644)
 }
 
-func (p *SourcePlugin) jsonifyTables(tables schema.Tables) []jsonTable {
+func (p *Plugin) jsonifyTables(tables schema.Tables) []jsonTable {
 	jsonTables := make([]jsonTable, len(tables))
 	for i, table := range tables {
 		jsonColumns := make([]jsonColumn, len(table.Columns))
@@ -77,7 +83,7 @@ func (p *SourcePlugin) jsonifyTables(tables schema.Tables) []jsonTable {
 	return jsonTables
 }
 
-func (p *SourcePlugin) renderTablesAsMarkdown(dir string) error {
+func (p *Plugin) renderTablesAsMarkdown(dir string) error {
 	for _, table := range p.Tables() {
 		if err := renderAllTables(table, dir); err != nil {
 			return err
@@ -90,15 +96,17 @@ func (p *SourcePlugin) renderTablesAsMarkdown(dir string) error {
 		return fmt.Errorf("failed to parse template for README.md: %v", err)
 	}
 
+	var b bytes.Buffer
+	if err := t.Execute(&b, p); err != nil {
+		return fmt.Errorf("failed to execute template: %v", err)
+	}
+	content := formatMarkdown(b.String())
 	outputPath := filepath.Join(dir, "README.md")
 	f, err := os.Create(outputPath)
 	if err != nil {
 		return fmt.Errorf("failed to create file %v: %v", outputPath, err)
 	}
-	defer f.Close()
-	if err := t.Execute(f, p); err != nil {
-		return fmt.Errorf("failed to execute template: %v", err)
-	}
+	f.WriteString(content)
 	return nil
 }
 
@@ -115,7 +123,7 @@ func renderAllTables(t *schema.Table, dir string) error {
 }
 
 func renderTable(table *schema.Table, dir string) error {
-	t := template.New("").Funcs(map[string]interface{}{
+	t := template.New("").Funcs(map[string]any{
 		"formatType": formatType,
 	})
 	t, err := t.New("table.md.go.tpl").ParseFS(templatesFS, "templates/table.md.go.tpl")
@@ -124,15 +132,23 @@ func renderTable(table *schema.Table, dir string) error {
 	}
 
 	outputPath := filepath.Join(dir, fmt.Sprintf("%s.md", table.Name))
+
+	var b bytes.Buffer
+	if err := t.Execute(&b, table); err != nil {
+		return fmt.Errorf("failed to execute template: %v", err)
+	}
+	content := formatMarkdown(b.String())
 	f, err := os.Create(outputPath)
 	if err != nil {
 		return fmt.Errorf("failed to create file %v: %v", outputPath, err)
 	}
-	defer f.Close()
-	if err := t.Execute(f, table); err != nil {
-		return fmt.Errorf("failed to execute template: %v", err)
-	}
-	return nil
+	f.WriteString(content)
+	return f.Close()
+}
+
+func formatMarkdown(s string) string {
+	s = reMatchNewlines.ReplaceAllString(s, "\n\n")
+	return reMatchHeaders.ReplaceAllString(s, `$1`+"\n\n")
 }
 
 func formatType(v schema.ValueType) string {