+++ plugins/source/metrics.go
@@ -1,4 +1,4 @@
-package plugins
+package source
 
 import (
 	"time"
@@ -6,7 +6,7 @@ import (
 	"github.com/cloudquery/plugin-sdk/schema"
 )
 
-type SourceMetrics struct {
+type Metrics struct {
 	TableClient map[string]map[string]*TableClientMetrics
 }
 
@@ -23,7 +23,7 @@ func (s *TableClientMetrics) Equal(other *TableClientMetrics) bool {
 }
 
 // Equal compares to stats. Mostly useful in testing
-func (s *SourceMetrics) Equal(other *SourceMetrics) bool {
+func (s *Metrics) Equal(other *Metrics) bool {
 	for table, clientStats := range s.TableClient {
 		for client, stats := range clientStats {
 			if _, ok := other.TableClient[table]; !ok {
@@ -53,27 +53,17 @@ func (s *SourceMetrics) Equal(other *SourceMetrics) bool {
 	return true
 }
 
-func (s *SourceMetrics) initWithTables(tables schema.Tables) {
-	for _, table := range tables {
-		if _, ok := s.TableClient[table.Name]; !ok {
-			s.TableClient[table.Name] = make(map[string]*TableClientMetrics)
-		}
-		s.initWithTables(table.Relations)
-	}
-}
-
-func (s *SourceMetrics) initWithClients(table *schema.Table, clients []schema.ClientMeta) {
+func (s *Metrics) initWithClients(table *schema.Table, clients []schema.ClientMeta) {
+	s.TableClient[table.Name] = make(map[string]*TableClientMetrics, len(clients))
 	for _, client := range clients {
-		if _, ok := s.TableClient[table.Name][client.ID()]; !ok {
-			s.TableClient[table.Name][client.ID()] = &TableClientMetrics{}
-		}
-		for _, relation := range table.Relations {
-			s.initWithClients(relation, clients)
-		}
+		s.TableClient[table.Name][client.ID()] = &TableClientMetrics{}
+	}
+	for _, relation := range table.Relations {
+		s.initWithClients(relation, clients)
 	}
 }
 
-func (s *SourceMetrics) TotalErrors() uint64 {
+func (s *Metrics) TotalErrors() uint64 {
 	var total uint64
 	for _, clientMetrics := range s.TableClient {
 		for _, metrics := range clientMetrics {
@@ -83,7 +73,7 @@ func (s *SourceMetrics) TotalErrors() uint64 {
 	return total
 }
 
-func (s *SourceMetrics) TotalPanics() uint64 {
+func (s *Metrics) TotalPanics() uint64 {
 	var total uint64
 	for _, clientMetrics := range s.TableClient {
 		for _, metrics := range clientMetrics {
@@ -93,7 +83,7 @@ func (s *SourceMetrics) TotalPanics() uint64 {
 	return total
 }
 
-func (s *SourceMetrics) TotalResources() uint64 {
+func (s *Metrics) TotalResources() uint64 {
 	var total uint64
 	for _, clientMetrics := range s.TableClient {
 		for _, metrics := range clientMetrics {