+++ plugins/source/scheduler_dfs.go
@@ -1,4 +1,4 @@
-package plugins
+package source
 
 import (
 	"context"
@@ -22,7 +22,7 @@ const (
 	minResourceConcurrency = 100
 )
 
-func (p *SourcePlugin) syncDfs(ctx context.Context, spec specs.Source, client schema.ClientMeta, tables schema.Tables, resolvedResources chan<- *schema.Resource) {
+func (p *Plugin) syncDfs(ctx context.Context, spec specs.Source, client schema.ClientMeta, tables schema.Tables, resolvedResources chan<- *schema.Resource) {
 	// This is very similar to the concurrent web crawler problem with some minor changes.
 	// We are using DFS to make sure memory usage is capped at O(h) where h is the height of the tree.
 	tableConcurrency := max(spec.Concurrency/minResourceConcurrency, minTableConcurrency)
@@ -36,19 +36,24 @@ func (p *SourcePlugin) syncDfs(ctx context.Context, spec specs.Source, client sc
 	}
 	p.resourceSem = semaphore.NewWeighted(int64(resourceConcurrency))
 
-	var wg sync.WaitGroup
-	for _, table := range tables {
-		if table.Parent != nil {
-			// skip descendent tables here - they will be handled by the recursive
-			// depth-first-search later.
-			continue
-		}
-		table := table
+	// we have this because plugins can return sometimes clients in a random way which will cause
+	// differences between this run and the next one.
+	preInitialisedClients := make([][]schema.ClientMeta, len(tables))
+	for i, table := range tables {
 		clients := []schema.ClientMeta{client}
 		if table.Multiplex != nil {
 			clients = table.Multiplex(client)
 		}
+		preInitialisedClients[i] = clients
+		// we do this here to avoid locks so we initial the metrics structure once in the main goroutines
+		// and then we can just read from it in the other goroutines concurrently given we are not writing to it.
 		p.metrics.initWithClients(table, clients)
+	}
+
+	var wg sync.WaitGroup
+	for i, table := range tables {
+		table := table
+		clients := preInitialisedClients[i]
 		for _, client := range clients {
 			client := client
 			if err := p.tableSems[0].Acquire(ctx, 1); err != nil {
@@ -62,23 +67,32 @@ func (p *SourcePlugin) syncDfs(ctx context.Context, spec specs.Source, client sc
 				defer p.tableSems[0].Release(1)
 				// not checking for error here as nothing much todo.
 				// the error is logged and this happens when context is cancelled
-				p.resolveTableDfs(ctx, tables, table, client, nil, resolvedResources, 0)
+				p.resolveTableDfs(ctx, table, client, nil, resolvedResources, 1)
 			}()
 		}
 	}
 	wg.Wait()
 }
 
-func (p *SourcePlugin) resolveTableDfs(ctx context.Context, allIncludedTables schema.Tables, table *schema.Table, client schema.ClientMeta, parent *schema.Resource, resolvedResources chan<- *schema.Resource, depth int) {
+func (p *Plugin) logTablesMetrics(tables schema.Tables, client schema.ClientMeta) {
+	clientName := client.ID()
+	for _, table := range tables {
+		metrics := p.metrics.TableClient[table.Name][clientName]
+		p.logger.Info().Str("table", table.Name).Str("client", clientName).Uint64("resources", metrics.Resources).Uint64("errors", metrics.Errors).Msg("table sync finished")
+		p.logTablesMetrics(table.Relations, client)
+	}
+}
+
+func (p *Plugin) resolveTableDfs(ctx context.Context, table *schema.Table, client schema.ClientMeta, parent *schema.Resource, resolvedResources chan<- *schema.Resource, depth int) {
 	clientName := client.ID()
 	logger := p.logger.With().Str("table", table.Name).Str("client", clientName).Logger()
 
 	if parent == nil { // Log only for root tables, otherwise we spam too much.
-		logger.Info().Msg("table resolver started")
+		logger.Info().Msg("top level table resolver started")
 	}
-
 	tableMetrics := p.metrics.TableClient[table.Name][clientName]
-	res := make(chan interface{})
+
+	res := make(chan any)
 	go func() {
 		defer func() {
 			if err := recover(); err != nil {
@@ -87,7 +101,7 @@ func (p *SourcePlugin) resolveTableDfs(ctx context.Context, allIncludedTables sc
 					scope.SetTag("table", table.Name)
 					sentry.CurrentHub().CaptureMessage(stack)
 				})
-				p.logger.Error().Interface("error", err).Str("stack", stack).Msg("table resolver finished with panic")
+				logger.Error().Interface("error", err).Str("stack", stack).Msg("table resolver finished with panic")
 				atomic.AddUint64(&tableMetrics.Panics, 1)
 			}
 			close(res)
@@ -100,16 +114,17 @@ func (p *SourcePlugin) resolveTableDfs(ctx context.Context, allIncludedTables sc
 	}()
 
 	for r := range res {
-		p.resolveResourcesDfs(ctx, allIncludedTables, table, client, parent, r, resolvedResources, depth)
+		p.resolveResourcesDfs(ctx, table, client, parent, r, resolvedResources, depth)
 	}
 
 	// we don't need any waitgroups here because we are waiting for the channel to close
-	if parent == nil { // Log only for root tables, otherwise we spam too much.
-		logger.Info().Uint64("resources", tableMetrics.Resources).Uint64("errors", tableMetrics.Errors).Msg("fetch table finished")
+	if parent == nil { // Log only for root tables and relations only after resolving is done, otherwise we spam per object instead of per table.
+		logger.Info().Uint64("resources", tableMetrics.Resources).Uint64("errors", tableMetrics.Errors).Msg("table sync finished")
+		p.logTablesMetrics(table.Relations, client)
 	}
 }
 
-func (p *SourcePlugin) resolveResourcesDfs(ctx context.Context, allIncludedTables schema.Tables, table *schema.Table, client schema.ClientMeta, parent *schema.Resource, resources interface{}, resolvedResources chan<- *schema.Resource, depth int) {
+func (p *Plugin) resolveResourcesDfs(ctx context.Context, table *schema.Table, client schema.ClientMeta, parent *schema.Resource, resources any, resolvedResources chan<- *schema.Resource, depth int) {
 	resourcesSlice := helpers.InterfaceSlice(resources)
 	if len(resourcesSlice) == 0 {
 		return
@@ -147,11 +162,6 @@ func (p *SourcePlugin) resolveResourcesDfs(ctx context.Context, allIncludedTable
 		resolvedResources <- resource
 		for _, relation := range resource.Table.Relations {
 			relation := relation
-			if allIncludedTables.GetTopLevel(relation.Name) == nil {
-				// this indicates that child table is skipped by user config,
-				// so we should not sync it
-				continue
-			}
 			if err := p.tableSems[depth].Acquire(ctx, 1); err != nil {
 				// This means context was cancelled
 				wg.Wait()
@@ -161,14 +171,14 @@ func (p *SourcePlugin) resolveResourcesDfs(ctx context.Context, allIncludedTable
 			go func() {
 				defer wg.Done()
 				defer p.tableSems[depth].Release(1)
-				p.resolveTableDfs(ctx, allIncludedTables, relation, client, resource, resolvedResources, depth+1)
+				p.resolveTableDfs(ctx, relation, client, resource, resolvedResources, depth+1)
 			}()
 		}
 	}
 	wg.Wait()
 }
 
-func (p *SourcePlugin) resolveResource(ctx context.Context, table *schema.Table, client schema.ClientMeta, parent *schema.Resource, item interface{}) *schema.Resource {
+func (p *Plugin) resolveResource(ctx context.Context, table *schema.Table, client schema.ClientMeta, parent *schema.Resource, item any) *schema.Resource {
 	ctx, cancel := context.WithTimeout(ctx, 10*time.Minute)
 	defer cancel()
 	resource := schema.NewResourceData(table, parent, item)
@@ -205,7 +215,7 @@ func (p *SourcePlugin) resolveResource(ctx context.Context, table *schema.Table,
 	return resource
 }
 
-func (p *SourcePlugin) resolveColumn(ctx context.Context, logger zerolog.Logger, tableMetrics *TableClientMetrics, client schema.ClientMeta, resource *schema.Resource, c schema.Column) {
+func (p *Plugin) resolveColumn(ctx context.Context, logger zerolog.Logger, tableMetrics *TableClientMetrics, client schema.ClientMeta, resource *schema.Resource, c schema.Column) {
 	columnStartTime := time.Now()
 	defer func() {
 		if err := recover(); err != nil {