+++ schema/json_test.go
@@ -10,11 +10,18 @@ type Foo struct {
 
 func TestJSONSet(t *testing.T) {
 	successfulTests := []struct {
-		source interface{}
+		source any
 		result JSON
 	}{
+		{source: "", result: JSON{Bytes: []byte(""), Status: Null}},
 		{source: "{}", result: JSON{Bytes: []byte("{}"), Status: Present}},
+		{source: `"test"`, result: JSON{Bytes: []byte(`"test"`), Status: Present}},
+		{source: "1", result: JSON{Bytes: []byte("1"), Status: Present}},
+		{source: "[1, 2, 3]", result: JSON{Bytes: []byte("[1, 2, 3]"), Status: Present}},
 		{source: []byte("{}"), result: JSON{Bytes: []byte("{}"), Status: Present}},
+		{source: []byte(`"test"`), result: JSON{Bytes: []byte(`"test"`), Status: Present}},
+		{source: []byte("1"), result: JSON{Bytes: []byte("1"), Status: Present}},
+		{source: []byte("[1, 2, 3]"), result: JSON{Bytes: []byte("[1, 2, 3]"), Status: Present}},
 		{source: ([]byte)(nil), result: JSON{Status: Null}},
 		{source: (*string)(nil), result: JSON{Status: Null}},
 
@@ -25,9 +32,9 @@ func TestJSONSet(t *testing.T) {
 		{source: []Foo{}, result: JSON{Bytes: []byte(`[]`), Status: Present}},
 		{source: []Foo{{1}}, result: JSON{Bytes: []byte(`[{"Num":1}]`), Status: Present}},
 
-		{source: map[string]interface{}{"foo": "bar"}, result: JSON{Bytes: []byte(`{"foo":"bar"}`), Status: Present}},
-		{source: map[string]interface{}(nil), result: JSON{Bytes: []byte(`{}`), Status: Present}},
-		{source: map[string]interface{}{}, result: JSON{Bytes: []byte(`{}`), Status: Present}},
+		{source: map[string]any{"foo": "bar"}, result: JSON{Bytes: []byte(`{"foo":"bar"}`), Status: Present}},
+		{source: map[string]any(nil), result: JSON{Bytes: []byte(`{}`), Status: Present}},
+		{source: map[string]any{}, result: JSON{Bytes: []byte(`{}`), Status: Present}},
 		{source: map[string]string{"foo": "bar"}, result: JSON{Bytes: []byte(`{"foo":"bar"}`), Status: Present}},
 		{source: map[string]string(nil), result: JSON{Bytes: []byte(`{}`), Status: Present}},
 		{source: map[string]string{}, result: JSON{Bytes: []byte(`{}`), Status: Present}},