+++ schema/macaddr.go
@@ -7,7 +7,7 @@ import (
 )
 
 type MacaddrTransformer interface {
-	TransformMacaddr(*Macaddr) interface{}
+	TransformMacaddr(*Macaddr) any
 }
 
 type Macaddr struct {
@@ -38,13 +38,13 @@ func (dst *Macaddr) String() string {
 	}
 }
 
-func (dst *Macaddr) Set(src interface{}) error {
+func (dst *Macaddr) Set(src any) error {
 	if src == nil {
 		*dst = Macaddr{Status: Null}
 		return nil
 	}
 
-	if value, ok := src.(interface{ Get() interface{} }); ok {
+	if value, ok := src.(interface{ Get() any }); ok {
 		value2 := value.Get()
 		if value2 != value {
 			return dst.Set(value2)
@@ -84,7 +84,7 @@ func (dst *Macaddr) Set(src interface{}) error {
 	return nil
 }
 
-func (dst Macaddr) Get() interface{} {
+func (dst Macaddr) Get() any {
 	switch dst.Status {
 	case Present:
 		return dst.Addr